attempt to checkout the PR's branch
This commit adds some debugging (to print the full JSON github context vars) and attempts to checkout the head ref before attempting to do a diff. Hopefully this will fix issues of "object not found" because the diff is currently being done in the repo's default branch (main or master), but that won't work if the PR is being submitted to a different branch (eg 'dev') or, maybe, if the PR is being submitted from a different branch.

See also

 * https://stackoverflow.com/questions/70104600/complete-list-of-github-actions-contexts
 * https://github.com/maltfield/detect-malicious-unicode/issues/4
 * https://github.com/BusKill/buskill-app/pull/90#issuecomment-2274260749